Add option to not count weekends for age
Changes how Stale.since() counts days to add an option to not count
weekends against the age of an issue. When skipWeekends is true, we'll
go backward in time only counting weekdays until we get to the age. When
skipWeekends is false (the default) we'll shortcut this and treat the
argument to since() as a literal number of days.

Fixes #130